# Conflicts:
#	src/main/resources/com/app/flight/data/csv/Flight.csv
#	src/main/resources/com/app/flight/data/json/BoardingPass.json

+++ src/main/java/com/app/flight/entity/Seat.java
@@ -9,7 +9,7 @@ public enum Seat {
     private double price;
     private int row;
 
-    Seat(int price, int row) {
+    Seat(double price, int row) {
         this.price = price;
         this.row = row;
     }
